+++ tests/contrib/test_securecookie.py
@@ -14,7 +14,7 @@ import pytest
 
 from werkzeug._compat import to_native
 from werkzeug.contrib.securecookie import SecureCookie
-from werkzeug.utils import parse_cookie
+from werkzeug.http import parse_cookie
 from werkzeug.wrappers import Request
 from werkzeug.wrappers import Response